Fix Issue #1 Moved all header files to the include folder in src and modified the CMAKELists.txt file to reflect those changes.
