+++ src/database.cpp
@@ -1,4 +1,4 @@
-#include "database.hpp"
+#include "src/include/database.hpp"
 
 database* database::instance;
 std::atomic<database*> database::s_instance;
