+++ src/include/inference_worker.hpp
@@ -3,7 +3,7 @@
 
 #include <string>
 
-#include "callback_timer.hpp"
+#include "src/include/callback_timer.hpp"
 
 class inference_worker
 {
