+++ src/mainwindow.cpp
@@ -1,4 +1,4 @@
-#include "mainwindow.h"
+#include "src/include/mainwindow.h"
 #include "ui_mainwindow.h"
 
 #include <pcap.h>
