+++ src/packet_capture.cpp
@@ -1,4 +1,4 @@
-#include "packet_capture.hpp"
+#include "src/include/packet_capture.hpp"
 
 packet_capture::packet_capture() {
 	int num_packets = 30;
