+++ src/worker/moc_ui_reload_worker.cpp
@@ -6,7 +6,7 @@
 ** WARNING! All changes made in this file will be lost!
 *****************************************************************************/
 
-#include "ui_reload_worker.hpp"
+#include "src/include/ui_reload_worker.hpp"
 #include <QtCore/qbytearray.h>
 #include <QtCore/qmetatype.h>
 #if !defined(Q_MOC_OUTPUT_REVISION)
