+++ src/worker/packet_capture_worker.cpp
@@ -1,7 +1,7 @@
-#include "packet_capture_worker.h"
+#include "src/include/packet_capture_worker.h"
 
-#include "../flow_generator_engine.hpp"
-#include "ui_reload_worker.hpp"
+#include "src/include/flow_generator_engine.hpp"
+#include "src/include/ui_reload_worker.hpp"
 
 packet_capture_worker::packet_capture_worker(std::vector<basic_packet_info>& buff,
 											 char* filter,
